solve errors

+++ src/app/shopping-card-and-purchase/item-card/item-card.component.ts
@@ -1,9 +1,11 @@
 import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { Book } from 'src/app/shared/book.model';
 import { Order } from 'src/app/shared/order.model';
+import { AuthService } from 'src/app/shared/services/auth.service';
 import { BookService } from 'src/app/shared/services/books.service';
 import { OrderPaymentService } from 'src/app/shared/services/orderPayment.service';
 import { ShCardService } from 'src/app/shared/services/shcard.service';
+import { UIService } from 'src/app/shared/services/ui.service';
 
 @Component({
   selector: 'app-item-card',
@@ -14,10 +16,13 @@ export class ItemCardComponent implements OnInit {
   @Input() order: Order;
   book: Book;
   needSpinner = true;
+  auth = false;
   constructor(
     private bookService: BookService,
     private orderPayService: OrderPaymentService,
-    private shCardService: ShCardService
+    private shCardService: ShCardService,
+    private authService: AuthService,
+    private uiService: UIService
   ) {}
 
   ngOnInit() {
@@ -30,11 +35,18 @@ export class ItemCardComponent implements OnInit {
       // console.log('aBook:Book item order ', aBook);
       this.needSpinner = false;
     });
+    this.authService.isAuth.subscribe((a) => {
+      this.auth = a;
+    });
   }
   payThisOrder(count: any) {
-    console.log(this.order);
-    this.orderPayService.payOrder(this.order.bookId, count);
-    this.shCardService.removeFromCard(this.order.bookId);
+    if (this.auth) {
+      // console.log(this.order);
+      this.orderPayService.payOrder(this.order.bookId, count);
+      this.shCardService.removeFromCard(this.order.bookId);
+    } else {
+      this.uiService.presentToast('لطفا وارد شوید ');
+    }
   }
   deleteThisOrder() {
     console.log(this.book.id);
